Use Systemd for Amazon Linux 2 packages

+++ scripts/post-uninstall.sh
@@ -43,11 +43,15 @@ elif [[ -f /etc/lsb-release ]]; then
     fi
 elif [[ -f /etc/os-release ]]; then
     source /etc/os-release
-    if [[ $ID = "amzn" ]]; then
-        # Amazon Linux logic
-        if [[ "$1" = "0" ]]; then
-            # InfluxDB is no longer installed, remove from init system
-            rm -f /etc/default/influxdb
+    if [[ "$ID" = "amzn" ]] && [[ "$1" = "0" ]]; then
+        # InfluxDB is no longer installed, remove from init system
+        rm -f /etc/default/influxdb
+
+        if [[ "$NAME" = "Amazon Linux" ]]; then
+            # Amazon Linux 2+ logic
+            disable_systemd
+        elif [[ "$NAME" = "Amazon Linux AMI" ]]; then
+            # Amazon Linux logic
             disable_chkconfig
         fi
     fi
